refactor: separate query factories by operations
- Created 'factory' package
- Moved query factories from 'util' to 'factory' package
- Separated Transaction and User query factories by SQL operations
