+++ src/main/java/org/example/finlog/factory/transaction/query/TransactionQueryFactory.java
@@ -1,0 +1,7 @@
+package org.example.finlog.factory.transaction.query;
+
+import org.example.finlog.util.TableName;
+
+public abstract class TransactionQueryFactory {
+    protected static final String TABLE = TableName.TRANSACTION;
+}
