+++ src/main/java/org/example/finlog/factory/user/query/UserDeleteFactory.java
@@ -1,0 +1,4 @@
+package org.example.finlog.factory.user.query;
+
+public class UserDeleteFactory extends UserQueryFactory {
+}
