+++ src/main/java/org/example/finlog/factory/user/query/UserQueryFactory.java
@@ -1,0 +1,7 @@
+package org.example.finlog.factory.user.query;
+
+import org.example.finlog.util.TableName;
+
+public abstract class UserQueryFactory {
+    protected final static String TABLE = TableName.USER;
+}
